requirements.txt
dashscope
speechrecognition
pyaudio
flask
python-dotenv